hide some of the boilerplate in setup
