+++ waspc/data/Generator/templates/react-app/vite.config.ts
@@ -15,6 +15,7 @@ export default defineConfig({
   },
   test: {
     globals: true,
-    environment: 'jsdom'
+    environment: 'jsdom',
+    setupFiles: ['./vitest.setup.tsx'],
   },
 })
